soong: add the sbox_lineage command
This is a copy of the sbox command from the build/soong repo from
Lineage. Compared to the stock sbox from Google, it adds the
parameter '--copy-all-files' that is used to build the kernel headers.

We add this as new command to avoid editing the AOSP soong and blueprint
projects.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>